requirements.txt
opencv-python
numpy
pyserial
PyYAML
picamera2
pytest